+++ package.json
@@ -18,6 +18,7 @@
     "framer-motion": "^8.5.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.43.1",
     "react-icons": "^4.7.1",
     "react-multi-date-picker": "^3.3.4",
     "react-redux": "^8.0.5",
