+++ src/pages/HomePage.tsx
@@ -1,5 +1,5 @@
 //TODO:  navbar, search bar, pet sitters list and map
-import SearchBar from "../components/SearchBar";
+import SearchBar from "../components/SearchBar/SearchBar";
 import PetSitterCard from "../components/PetSitterCard";
 import GoogleMap from "../components/GoogleMap";
